third_party/blink/web_tests/fast/forms/text/text-selection-after-type-change.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<input>

<script>
test(() => {
  assert_own_property(window, 'eventSender');
  var element = document.querySelector('input');
  element.value = '';
  element.focus();
  eventSender.keyDown('c');
  assert_equals(element.selectionStart, 1);
  element.type = 'button';
  assert_equals(element.selectionStart, null);
  element.type = 'text';

  // TODO(tkent): We need some delay or a force layout before checking
  // selectionStart because input type change doesn't update selectionStart
  // immediately to avoid force layout. crbug.com/721217
  element.offsetWidth;
  assert_equals(element.selectionStart, 1);
}, 'selectionStart/selectionEnd should be kept after input type changes.');
</script>
</body>
